+++ lib/df_http_client_config.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:math';
 
 import '/models/result.dart';
 
@@ -19,6 +20,11 @@ class DfHttpClientConfig {
   /// - [waitForTokenRefresh]: Whether requests should wait until token refresh completes.
   /// - [maxRetryAttempts]: Maximum number of retry attempts for failed requests.
   /// - [refreshToken]: Callback used to refresh tokens if authentication fails.
+  /// - [maxDelayMs]: Represent the maximum waiting time on API retry. Default set to _60000ms_
+  /// - [internetConnectionCheck] Check if the device is really connected to the internet by default it
+  /// check if the 'example.com' is available, the custom logic can be implemented
+  /// ---
+  /// - [random]: Provide a "seeded" Random in tests so that the retry delay is predictable
   ///
   DfHttpClientConfig({
     required this.baseApiUrl,
@@ -29,8 +35,14 @@ class DfHttpClientConfig {
     this.waitForTokenRefresh = true,
     this.maxRetryAttempts = 3,
     this.refreshToken,
-  });
+    this.maxDelayMs = 60000,
+    Random? random,
+    Future<bool> Function()? internetConnectionCheck,
+  })
+      : _rand = random ?? Random(),
+        hasInternetConnection = internetConnectionCheck??_hasInternetConnection;
 
+  final int maxDelayMs;
   final String baseApiUrl;
   final Encoding? encoding;
   Map<String, String> headers;
@@ -39,6 +51,11 @@ class DfHttpClientConfig {
   bool waitForTokenRefresh;
   final int maxRetryAttempts;
   Future<Result<String, Exception>> Function()? refreshToken;
+
+  Future<bool> Function() hasInternetConnection;
+
+  /// Used to generate random jitter value for the retry waiting time
+  final Random _rand;
 
   void addHeaderParameters(Map<String, String> headers) {
     this.headers.addAll(headers);
@@ -65,7 +82,7 @@ class DfHttpClientConfig {
     if (authorizationToken != null) {
       authorizationToken = authorizationToken.replaceFirst("Bearer ", "");
     }
-    return headers[HttpHeaders.authorizationHeader];
+    return authorizationToken;
   }
 
   DfHttpClientConfig clone() {
@@ -100,5 +117,36 @@ class DfHttpClientConfig {
       timeout: timeout ?? this.timeout,
       waitForTokenRefresh: waitForTokenRefresh ?? this.waitForTokenRefresh,
     );
+  }
+
+  int calculateRetryWaitingPeriod(int retryCount) {
+    // Exponential backoff with jitter
+    final attemptsUsed = (maxRetryAttempts - retryCount);
+    final baseMs = 500; // base delay in ms
+    final cappedAttempts = min(attemptsUsed, 10); // prevents huge shifts
+    final exponential =
+        baseMs * (1 << cappedAttempts); // base * 2^cappedAttempts
+    final jitter = _rand.nextInt(200); // 0..199 ms jitter
+    var retryPauseDurationMs = exponential + jitter;
+
+    //Prevents too long retry awaits
+    if (retryPauseDurationMs > maxDelayMs) {
+      retryPauseDurationMs = maxDelayMs;
+    }
+
+    return retryPauseDurationMs;
+  }
+
+  /// Checks whether the device currently has an active internet connection.
+  ///
+  /// Returns `true` if DNS lookup succeeds, otherwise `false`.
+  static Future<bool> _hasInternetConnection() async {
+    try {
+      final result = await InternetAddress.lookup('example.com');
+
+      return result.isNotEmpty && result[0].rawAddress.isNotEmpty;
+    } catch (_) {
+      return false;
+    }
   }
 }